Fix `mac install vim` to show an error if a dep install fails
Previously a dep would look like it had installed successfully
but had actually failed silently leading to a great deal of
confusion. This fixes it so that if there is a failure installing
the dep it will show the output from the failed command and then
return an exit code of 1.

This kind of pattern of capturing output and then checking the exit
code seems like something that could be more generally useful. I may
look at extracting a more general solution in the future.

Also this tweaks `mac install` so that it stops on failure and
returns an appropriate error code.