Drop pathogen in favour of vim 8 standard plugins
Turns out there was some residual config on my machine that was making
things continue to work even though the dotmac vim config was missing
some important pieces. This fixes it by ditching pathongen and adding
the parts that were missing. I opted to ditch pathogen as getting it
to play nice with my .vim directory not being at ~/.vim was difficult.
It's probably possible to make it work, but as vim 8 plugins also work
just fine it seems a better solution anyways. This also adds the
jellybeans theme in which was missing.