Don't attempt to load chruby the first time
Attempting to do this before chruby is properly setup breaks the
installation, which prevents chruby from being properly setup, which
breaks the installation...

Dropping the attempt to invoke chruby from the install script fixes
this at the price of not ensuring the correct ruby version is selected
for the current terminal. Eh terminal sessions are cheap.