Add nvm for managing node versions
Since nvm is super slow to run their setup script I've opted not
to source it as part of my default shell setup as they suggest and
I've instead created an `nvm` function that when invoked sources
the nvm.sh file and then executes the desired nvm command. That way
I only pay the cost of it when running nvm the first time in a shell
but for shells that don't need it they aren't delayed by ~300ms to
startup needlessly.