Add `git-mtm` for moving from the current code to the master branch
Once I push up code in a PR I often keep working on that branch,
sometimes because of changes I'm making being depdendent on the code
in the PR, othertimes purely due to laziness or forgetfulness. This
leads me to often going through the process of stashing my changes,
switching back to master, pulling the latest changes, and then popping
the stashed changes. It's not too much of a pain, but I do it enough
that I've finally written a very quick script to cover it so that
I can just write `git mtm` instead of having to go through all those
steps myself each time.