Install goimports using 'go install' instead of 'go get'
This fixes the installation of the `goimports` command to use the appropriate
`go install` command instead of the `go get` command which is no longer used
for this for quite awhile now.

Additionally this also switches the location of the default go directory to
"$MAC_HOME/go" instead of defaulting to "~/go", as I prefer to have more of my
files stored in the mac directory.