Don't commit Brewfile.lock.json files
Not sure if this is the right call yet, I may revert this later if I decide I
want to keep them as they could be helpful for making setup more reproducible,
but the current approach of erring towards just grabbing the latest versions
of things also hasn't bitten me too severly yet so I'm going to keep rolling
with it. For now this just ensures my git status is a bit cleaner.