Fix bug where vim package was symlinking inside itself
On the first run of mac install vim it would work fine, but running it again
afterwards would result in symlink being created to the vim package inside
itself. It doesn't really cause any problems, it just shouldn't be there and
clutters up git status. This prevents that from happening.